vbox-machine.inc: update PREFERRED_VERSION_linux-yocto 3.14 -> 4.1
The 3.14 kernel has been removed from the post jethro oe-core master
branch, so update to 4.1 (which is available in both jethro and
current oe-core master).

Note that the 4.4 kernel recently added to oe-core master causes
problems for vboxguestdrivers 4.3.30 (the current version in meta-oe)
so keep a preferred version for linux-yocto instead of just taking the
latest available version.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>